==> 06_roadapi/01_roadapi.ipynb <==
#행정안전부 도로명 API

건문명으로 도로명주소 구하기

import requests
from urllib.parse import urlparse,quote
import pandas as pd 

key_fd = open('roadapikey.txt' , mode ='r')
roadapi_key = key_fd.read(100)
key_fd.close()

road_url = 'https://www.juso.go.kr/addrlink/addrLinkApi.do'
url = f'{road_url}?confmKey={roadapi_key}&currentPage=1&countPerPage=10&keyword=강서구청&resultType=json'
urlparse(url).geturl()

quote('강서구청')

url = f'{road_url}?confmKey={roadapi_key}&currentPage=1&countPerPage=10&keyword={quote("강서구청")}&resultType=json'
url

results = requests.get(urlparse(url).geturl()).json()       
results

addr_list = results['results']['juso']
len(addr_list)

for addr in addr_list:
    print(addr['bdNm'],addr['roadAddr'])

예제> 5군데 구하기

addr_list = []; a_list=['강서구청','양천구청','부천시청','영등포구청','김포시청']
for a in a_list:
    road_url = 'https://www.juso.go.kr/addrlink/addrLinkApi.do'
    url = f'{road_url}?confmKey={roadapi_key}&currentPage=1&countPerPage=10&keyword={quote(a)}&resultType=json'
    urlparse(url).geturl(url)

    results = requests.get(urlparse(url).geturl()).json()
    addr = results['results']['juso'][0]['roadAddr']
   
    addr_list.append(addr)


addr_list

road_list = pd.DataFrame({
    '장소명' : a_list,
    '주소' : addr_list
})
road_list

road_list.to_csv('address.csv',sep=',',encoding='utf8', index=False)
