==> 06_roadapi/02_kakaoapi.ipynb <==
import requests
from urllib.parse import urlparse
import pandas as pd 

key_fd = open('kakaomapkey.txt' , mode ='r')
kmap_key = key_fd.read(100)
key_fd.close()

addr = '서울특별시 강서구 화곡로 302 (화곡동)'
url ='https://dapi.kakao.com/v2/local/search/address.json?query=' + addr

result = requests.get(urlparse(url).geturl(), headers ={"Authorization" : "KakaoAK "+kmap_key}).json()
result

result['documents'][0]['address_name'] 

result['documents'][0]['x'] 

result['documents'][0]['y'] 

lat = float(result['documents'][0]['x'] )
lng = float(result['documents'][0]['y'] )
lat,lng

address.csv 파일 읽어서 위도 경도 정보 추가하기

df =pd.read_csv('address.csv')
df

df.index

lat_list = []; lng_list=[]
for i in df.index:
    url ='https://dapi.kakao.com/v2/local/search/address.json?query=' + df['주소'][i]

    result = requests.get(urlparse(url).geturl(), headers ={"Authorization" : "KakaoAK "+kmap_key}).json()
    lat = float(result['documents'][0]['y'] )
    lng = float(result['documents'][0]['x'] )
    lat_list.append(lat)
    lng_list.append(lng)

df['lat'] = lat_list 
df['lng'] = lng_list
df

df.to_csv('address2.csv',sep=',',encoding='utf8', index=False)
