==> 06_roadapi/04_weather.ipynb <==
import requests
from urllib.parse import urlparse
import pandas as pd

key_fd = open('openweatherkey.txt', mode='r')
oweather_key = key_fd.read(100)
key_fd.close()


lat = 37.550966
lng = 126.849532
url = f'http://api.openweathermap.org/data/2.5/weather?lat={lat}&lon={lng}&appid={oweather_key}&units=metric&lang=kr'

result = requests.get(urlparse(url).geturl()).json()
result

main = result['weather'][0]['main']
desc = result['weather'][0]['description']
main, desc


temp = result['main']['temp']
temp_min = result['main']['temp_min']
temp_max = result['main']['temp_max']
temp, temp_min, temp_max

temp = round(float(temp)+0.01, 1)
temp

df = pd.read_csv('address2.csv')
df


weather_list = []; desc_list = []; temp_list = []
for i in df.index:
    url = f'http://api.openweathermap.org/data/2.5/weather?lat={df.lat[i]}&lon={df.lng[i]}&appid={oweather_key}&units=metric&lang=kr'
    result = requests.get(urlparse(url).geturl()).json()
    main = result['weather'][0]['main']
    desc = result['weather'][0]['description']
    temp = result['main']['temp']
    temp = round(float(temp)+0.01, 1)
    temp_min = result['main']['temp_min']
    temp_max = result['main']['temp_max']
    weather_list.append(main)
    desc_list.append(desc)
    temp_list.append(f'{temp}/{temp_min}/{temp_max}')

df['weather'] = weather_list
df['desc'] = desc_list
df['temp'] = temp_list

df.to_csv('address3.csv', sep=',', encoding='utf8', index=False)
