==> 06_roadapi/05_folium.ipynb <==
folium 이용해서 지도 표시하기

import folium
import pandas as pd 

folium.__version__

mapping = folium.Map(location=[37.550966,126.849532], zoom_start=13)
mapping

mapping = folium.Map(location=[37.550966,126.849532],tiles='Stamen Toner', zoom_start=13)
mapping

mapping = folium.Map(
    location=[37.550966,126.849532],tiles='Stamen Toner', zoom_start=13)

folium.Marker(
    location=[37.517075, 126.866543],
    popup='양천구청',
    icon=folium.Icon(icon='cloud')
).add_to(mapping)

folium.Marker(
    location=[37.502277, 126.765004],
    popup='부천시청',
    icon=folium.Icon(color='green')
).add_to(mapping)

folium.Marker(
    location=[37.525963, 126.896367],
    popup='영등포구청',
    icon=folium.Icon(color='red', icon='info-sign')
).add_to(mapping)

folium.Marker(
    location=[37.615190, 126.715717],
    popup='김포시청',
    icon=folium.Icon(color='blue', icon='info-sign')
).add_to(mapping)

mapping

address3.csv 를 읽어서 지도위에 표시하기

df = pd.read_csv('address3.csv')
df

df['icon'] = ['map-marker','cloud','gift','info-sign','pk-circle']
df

color_dict = {'Clear' :'blue','Snow':'white', 'Rain':'grey','Extreme':'red','Clouds':'orange','Mist':'green'}

mapping = folium.Map(
    location=[df.lat.mean(),df.lng.mean()], zoom_start=12)



for i in df.index:
    folium.Marker(
    location=[df.lat[i], df.lng[i]],
    popup = df['장소명'][i],
    tooltip =f'{df.desc[i]}, {df.temp[i]}',
    icon=folium.Icon(color=color_dict[df.weather[i]], icon=df.icon[i])
).add_to(mapping)
mapping

마커 지도를 웹에 표시 > flask설치-> marker.py 만들어서 -> http://localhost:5000/marker 여기로 열기.
