==> 06_roadapi/09_stations1.ipynb <==
import numpy as np
import pandas as pd 
from glob import glob
import requests
from urllib.parse import quote
from tqdm import tqdm_notebook
import json
import folium
import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)

import time
from selenium import webdriver
from selenium.webdriver.common.keys import Keys

# 셀레니움 써서 해보기

# 엑셀 다운 
driver = webdriver.Chrome('../crawling/chromedriver')
driver.maximize_window()
driver.get('https://www.opinet.co.kr/user/main/mainView.do')
time.sleep(1)
driver.find_element_by_css_selector('.ic_m1').click()
time.sleep(2)
driver.find_element_by_xpath('//*[@id="SIGUNGU_NM0"]/option[5]').click()
time.sleep(1)



df = pd.read_excel('./gangseogu_oil.xls',header=2)
df

# 강서구 저렴한곳 10곳

top10 = df.sort_values(by='휘발유', ascending=True).head(10)
top10

top10_df = pd.DataFrame({
    '상호': top10['상호'],
    '주소': top10['주소'],
    '가격': top10['휘발유'],
    '셀프': top10['셀프여부'],
    '상표': top10['상표'],
    '전화번호' : top10['전화번호']
})
top10_df

key_fd = open('../crawling/kakaomapkey.txt', mode='r')
kmap_key = key_fd.read(100)
key_fd.close()
